# docs/openapi/lottomax-enterprise.yaml
openapi: 3.1.0
info:
  title: LottoMax Enterprise API
  version: 1.0.0
  summary: CIAM, wallet, lottery broad group, settlement, and admin risk APIs.
  description: >
    Production API contract for LottoMax real-money wallet and lottery operations.
    All monetary fields use integer minor units. Payment card data is never accepted
    by LottoMax APIs; gateway tokens and signed webhook references are used instead.
servers:
  - url: https://api.lottomax.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /admin/broad-groups:
    post:
      tags: [Admin, Lottery]
      operationId: createBroadGroup
      summary: Create a broad lottery group.
      parameters:
        - $ref: "#/components/parameters/AdminId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadGroupCreateRequest"
      responses:
        "201":
          description: Broad group created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadGroup"
...
  /draws/{drawId}/settlement:
    post:
      tags: [Settlement]
      operationId: settleDraw
      summary: Freeze pool, select winner, and post settlement.
      parameters:
        - $ref: "#/components/parameters/DrawId"
        - $ref: "#/components/parameters/AdminId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettleDrawRequest"
      responses:
        "200":
          description: Settlement posted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 or AWS Cognito JWT with MFA and KYC claims.
    gatewaySignature:
      type: apiKey
      in: header
      name: X-Gateway-Signature
  parameters:
...
    DrawId:
      name: drawId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AdminId:
      name: X-Admin-ID
      in: header
      required: true
      schema:
        type: string
...
  responses:
...
    Conflict:
      description: Conflict or duplicate idempotency key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Money:
      type: object
      required: [cents, currency]
      properties:
        cents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          enum: [USD, CAD]
...
    BroadGroupCreateRequest:
      type: object
      required: [name, maxPlayers, ticketPrice]
      properties:
        name:
          type: string
        maxPlayers:
          type: integer
          minimum: 2
        ticketPrice:
          $ref: "#/components/schemas/Money"
    BroadGroup:
      type: object
      required: [id, name, status, maxPlayers, ticketPrice]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [OPEN, PAUSED, FROZEN, SETTLED, VOID]
        maxPlayers:
          type: integer
        ticketPrice:
          $ref: "#/components/schemas/Money"
...
    SettleDrawRequest:
      type: object
      required: [rngNonce]
      properties:
        rngNonce:
          type: string
          minLength: 16
    Settlement:
      type: object
      required: [id, drawId, winnerUserId, winnerTicketId, companyAmount, winnerAmount, status]
      properties:
        id:
          type: string
          format: uuid
        drawId:
          type: string
          format: uuid
        winnerUserId:
          type: string
          format: uuid
        winnerTicketId:
          type: string
          format: uuid
        companyAmount:
          $ref: "#/components/schemas/Money"
        winnerAmount:
          $ref: "#/components/schemas/Money"
        status:
          type: string
          enum: [POSTED, REVERSED]
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string